requirements.txt
opencv-python-headless
pyzbar
Pillow
numpy
streamlit